% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.tools.R
\name{lp.tools}
\alias{lp.tools}
\alias{AddConstraint}
\title{Constraints}
\usage{
AddConstraint(description = "", mat, rhs, dir, avoid.duplicates = FALSE,
  rerun = TRUE, feedback = TRUE)
}
\arguments{
\item{description}{A description of the constraint.}
}
\value{
ToDo
}
\description{
Todo: add description

TODO.
}
\note{
%% ~~further notes~~
}
\examples{
require(lp.tools)
require(Matrix)
num.v <- 10
set.seed(60606)
Initialize(num.v)
SetObjective(objective = rep(1, num.v), description = "Sum", direction = "max")
AddConstraint(description = "Upper Bound",
              mat = Diagonal(num.v), 
              rhs = 1, 
              dir = "<=")
AddConstraint(description = "Random Bounds", 
              mat = matrix(runif(3 * num.v), ncol=num.v),
              rhs = 1, 
              dir = "<=")
AddConstraint(description = "Random Bound", 
              mat = runif(num.v),
              rhs = 1, 
              dir = "<=")
AddConstraint(description = "Random Bounds", 
              mat = Matrix(matrix(runif(3 * (num.v - 1)), ncol=(num.v - 1))),
              rhs = 1, 
              dir = "<=")
GetCostOfConstraint()

Initialize(num.v, lock.variables = rep(.3, 4))

SetObjective(objective = rep(1, num.v), description = "Sum", direction = "max")
AddConstraint(description = "Upper Bound", 
              mat = Diagonal(num.v), 
              rhs = 1, 
              dir = "<=")
AddConstraint(description = "Random Bounds", 
              mat = matrix(runif(3 * num.v), ncol=num.v),
              rhs = 1, 
              dir = "<=")

Initialize(num.v, lock.variables = rep(.3, 4))

SetObjective(objective.numer = runif(num.v), objective.denom  = runif(num.v), 
             description = "Fraction", direction = "max")
AddConstraint(description = "Upper Bound", 
              mat = Diagonal(num.v), 
              rhs = 1, 
              dir = "<=")
AddConstraint(description = "Random Bounds", 
              mat = matrix(runif(3 * num.v), ncol=num.v),
              rhs = 1, 
              dir = "<=")
RemoveConstraint(description = "Random Bounds")
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Georg
}
\keyword{~kwd1}
\keyword{~kwd2}
