% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.tools.R
\docType{class}
\name{Logger}
\alias{Logger}
\title{Logging via R6 object}
\format{\code{\link{R6Class}} object.}
\usage{
Logger
}
\value{
Object of \code{\link{R6Class}} Logger
}
\description{
Log to screen, file and to a data.frame.
}
\details{
Details: TODO.
}
\keyword{Log,}
\keyword{Logging}
