% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.tools.R
\docType{package}
\name{lp.tools-package}
\alias{lp.tools}
\alias{lp.tools-package}
\title{Tools to solve LP problems}
\description{
Helps formulating constraints and obectives for 
Linear Programming (LP) problems.
}

