% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.tools.R
\name{Other}
\alias{Other}
\alias{SetLPSolveScale}
\alias{Initialize}
\alias{Restore}
\alias{SetObjective}
\alias{RunLP}
\alias{GetThreeCols}
\alias{GetCostOfConstraint}
\alias{GetLastSolution}
\alias{RemoveConstraint}
\alias{CheckConstraints}
\alias{AddQuadConstraint}
\title{TODO: Divi up}
\usage{
SetLPSolveScale(opt.obj, lpsolve.scale)

Initialize(num.v, optimizer = "lpsolve", direction = "max", lock.variables,
  lpsolve.scale = 0, gurobi.params = list(), gurobi.output = "gurobi.txt",
  GetTimeStamp = function() paste(format(Sys.time(), "\%H-\%M-\%S"), ": ", sep
  = ""))

Restore(opt.obj)

SetObjective(opt.obj, objective, quad.obj, constant.numer = 0,
  objective.numer, constant.denom = 0, objective.denom, scale.factor = 1,
  direction, description = "value")

RunLP(opt.obj, dont.stop = FALSE)

GetThreeCols(mat)

GetCostOfConstraint(opt.obj, exclude = character(0))

GetLastSolution(opt.obj)

RemoveConstraint(opt.obj, description, rerun = TRUE, feedback = TRUE)

CheckConstraints(opt.objsolution, exclude = character(0), tolerance = 1e-06)

AddQuadConstraint(opt.obj, description = "", mat, rhs = 0, rerun = TRUE,
  feedback = TRUE)
}
\arguments{
\item{objective}{If this parameter is specified, then it defines the 
objective. If it is not sepcified, then the objective is fractional 
and is defined by the other four parameters.}
}
\description{
TODO.
}
