% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lp.tools.R
\name{lp.tools}
\alias{AddConstraint}
\alias{CheckConstraints}
\alias{GetCostOfConstraint}
\alias{GetLastSolution}
\alias{GetThreeCols}
\alias{GetValue}
\alias{Initialize}
\alias{RemoveConstraint}
\alias{Restore}
\alias{RunLP}
\alias{SetObjective}
\alias{lp.tools}
\title{Tools to solve LP problems}
\usage{
Initialize(num.v, optimizer, direction, lock.variables)

Restore()

AddConstraint(description = "", mat, rhs, dir)

SetObjective(objective, constant.numer = 0, objective.numer,
  constant.denom = 0, objective.denom, scale.factor = 1, direction,
  description = "value")

RunLP(dont.stop = FALSE)

GetValue(object.name)

GetThreeCols(mat)

GetCostOfConstraint(exclude = character(0))

GetLastSolution()

RemoveConstraint(description)

CheckConstraints(solution, exclude = character(0), tolerance = 1e-06)
}
\arguments{
\item{objective}{If this parameter is specified, then it defines the
objective. If it is not sepcified, then the objective is fractional
and is defined by the other four parameters.}

\item{grid.id}{A unique identifier for each grid tile.}

\item{lon}{A longitude value.}

\item{lat}{A latitiude value.}
}
\value{
ToDo
}
\description{
Helps formulating constraints and obectives for
Linear Programming (LP) problems.
}
\note{
%% ~~further notes~~
}
\examples{
require(lp.tools)
require(Matrix)
num.v <- 10
Initialize(num.v)
SetObjective(objective = rep(1, num.v), description = "Sum", direction = "max")
AddConstraint(description = "Upper Bound",
              mat = Diagonal(num.v),
              rhs = 1,
              dir = "<=")
AddConstraint(description = "Random Bounds",
              mat = matrix(runif(3 * num.v), ncol=num.v),
              rhs = 1,
              dir = "<=")
AddConstraint(description = "Random Bound",
              mat = runif(num.v),
              rhs = 1,
              dir = "<=")
GetCostOfConstraint()
}
\author{
Georg
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\keyword{~kwd1}
\keyword{~kwd2}

