% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lp.tools.R
\name{lp.tools}
\alias{AddConstraint}
\alias{GetCostOfConstraint}
\alias{GetValue}
\alias{Initialize}
\alias{RunLP}
\alias{SetObjective}
\alias{lp.tools}
\title{Tools to solve LP problems}
\usage{
Initialize(num.v, optimizer, direction)

AddConstraint(description = "", mat, rhs, dir)

SetObjective(objective, constant.numer = 0, objective.numer,
  constant.denom = 0, objective.denom, scale.factor = 1)

RunLP(dont.stop = FALSE)

GetValue(object.name)

GetCostOfConstraint()
}
\arguments{
\item{objective}{If this parameter is specified, then it defines the
objective. If it is not sepcified, then the objective is fractional
and is defined by the other four parameters.}

\item{grid.id}{A unique identifier for each grid tile.}

\item{lon}{A longitude value.}

\item{lat}{A latitiude value.}
}
\value{
ToDo
}
\description{
Helps formulating constraints and obectives for
Linear Programming (LP) problems.
}
\note{
%% ~~further notes~~
}
\examples{
4 + 5
}
\author{
Georg
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\keyword{~kwd1}
\keyword{~kwd2}

