% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.tools.R
\name{Other}
\alias{AddQuadConstraint}
\alias{CheckConstraints}
\alias{GetCostOfConstraint}
\alias{GetLastSolution}
\alias{GetThreeCols}
\alias{GetValue}
\alias{Initialize}
\alias{Other}
\alias{RemoveConstraint}
\alias{Restore}
\alias{RunLP}
\alias{SetLPSolveScale}
\alias{SetObjective}
\alias{SetValue}
\title{TODO: Divi up}
\usage{
SetLPSolveScale(lpsolve.scale)

Initialize(num.v, optimizer, direction, lock.variables)

Restore()

SetObjective(objective, constant.numer = 0, objective.numer,
  constant.denom = 0, objective.denom, scale.factor = 1, direction,
  description = "value")

RunLP(dont.stop = FALSE)

GetValue(object.name)

SetValue(object.name, value)

GetThreeCols(mat)

GetCostOfConstraint(exclude = character(0))

GetLastSolution()

RemoveConstraint(description, rerun = TRUE, feedback = TRUE)

CheckConstraints(solution, exclude = character(0), tolerance = 1e-06)

AddQuadConstraint(description = "", mat, rhs, dir, avoid.duplicates = FALSE,
  rerun = TRUE, feedback = TRUE)
}
\arguments{
\item{objective}{If this parameter is specified, then it defines the 
objective. If it is not sepcified, then the objective is fractional 
and is defined by the other four parameters.}
}
\description{
TODO.
}

